ML-API/requirements.txt
flask~=2.0.1
flask_wtf
requests
pandas~=1.2.5
sklearn~=0.0
matplotlib~=3.4.2
Pillow~=8.2.0
Pympler~=1.0.1
scikit-learn~=0.24.2
numpy~=1.21.0